+++ frontend/src/Styles/LongestSession.css
@@ -1,0 +1,231 @@
+:root {
+  --time: 10s;
+}
+
+.circle-container {
+  position: relative;
+  margin-top: 50px;
+  margin-left: 20px; /* Adjust the margin as needed */
+}
+
+.circle {
+  width: 300px;
+  height: 300px;
+  border-radius: 50%;
+  background-color: #ffffff; /* White background for the circle */
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  transition: transform 2s ease-in-out;
+  transform: scale(1);
+  border: 2px solid #2564eb00;
+  box-shadow: inset 0 0 10px rgba(0, 0, 0, 0.2), 0 10px 20px rgba(0, 0, 0, 0.2),
+    0 20px 30px rgba(0, 0, 0, 0.2), 0 30px 40px rgba(0, 0, 0, 0.2);
+}
+
+.circle:hover {
+  transform: scale(1.05);
+  box-shadow: inset 0 0 10px rgba(0, 0, 0, 0.3), 0 10px 20px rgba(0, 0, 0, 0.3),
+    0 20px 30px rgba(0, 0, 0, 0.3), 0 30px 40px rgba(0, 0, 0, 0.3);
+}
+
+.circle-content {
+  display: flex;
+  flex-direction: column;
+  text-align: center;
+}
+
+#time {
+  font-size: 30px;
+  margin: 0;
+  margin-left: 2em;
+  color: #2563eb;
+}
+
+.text {
+  margin-top: 4em;
+  font-size: 28px;
+  font-weight: 800;
+  text-align: center;
+}
+
+.timeContainer {
+  width: 50px;
+}
+
+#loader {
+  position: absolute;
+  top: 60%;
+  left: 35%;
+  transform: translate(-50%, -50%);
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  background-color: white;
+  padding: 0px;
+  margin-left: 0px;
+}
+
+.hourglass {
+  animation: flip var(--time) ease-in-out infinite;
+  border-bottom: solid 1vmin #630;
+  border-top: solid 1vmin #630;
+  left: 50%;
+  margin-left: -3vmin; /* Adjusted for smaller size */
+  margin-top: -5.5vmin; /* Adjusted for smaller size */
+  padding: 0 0.5vmin; /* Adjusted for smaller size */
+  position: fixed;
+  top: 50%;
+  transform: translate(-50%, -50%);
+}
+
+.top,
+.bottom {
+  background-color: #ffffff; /* White background to match the screen */
+  box-shadow: 0 0 1vmin 1vmin #bcd inset;
+  height: 5vmin; /* Adjusted for smaller size */
+  overflow: hidden;
+  position: relative;
+  width: 5vmin; /* Adjusted for smaller size */
+}
+
+.top {
+  border-radius: 0 0 50% 50%;
+}
+
+.top::before {
+  animation: top var(--time) linear infinite;
+  background-color: #000; /* Black sand */
+  border-radius: 50%;
+  content: "";
+  display: block;
+  height: 5vmin; /* Adjusted for smaller size */
+  left: 0;
+  position: absolute;
+  top: 0;
+  transform: translateY(50%);
+  width: 5vmin; /* Adjusted for smaller size */
+}
+
+.top::after {
+  animation: top-drip var(--time) linear infinite;
+  background-color: #000; /* Black sand */
+  content: "";
+  display: block;
+  height: 100%;
+  left: 45%;
+  position: absolute;
+  top: 0;
+  transform: translateY(100%);
+  width: 10%;
+}
+
+.bottom {
+  border-radius: 50% 50% 0 0;
+}
+
+.bottom::before {
+  animation: bottom var(--time) linear infinite;
+  background-color: #000; /* Black sand */
+  border-radius: 50%;
+  content: "";
+  display: block;
+  height: 5vmin; /* Adjusted for smaller size */
+  left: 0;
+  position: absolute;
+  top: 0;
+  transform: translateY(100%);
+  width: 5vmin; /* Adjusted for smaller size */
+}
+
+.bottom::after {
+  animation: bottom-drip var(--time) linear infinite;
+  background-color: #000; /* Black sand */
+  content: "";
+  display: block;
+  height: 100%;
+  left: 45%;
+  position: absolute;
+  top: 0;
+  width: 10%;
+}
+
+@keyframes flip {
+  0%,
+  45% {
+    transform: rotate(0);
+  }
+  50%,
+  95% {
+    transform: rotate(180deg);
+  }
+  100% {
+    transform: rotate(360deg);
+  }
+}
+
+@keyframes bottom {
+  0% {
+    transform: translateY(100%);
+  }
+  50% {
+    transform: translateY(50%);
+  }
+  51% {
+    transform: translateY(-50%);
+  }
+  100% {
+    transform: translateY(-100%);
+  }
+}
+
+@keyframes top {
+  0% {
+    transform: translateY(50%);
+  }
+  50% {
+    transform: translateY(100%);
+  }
+  51% {
+    transform: translateY(-100%);
+  }
+  100% {
+    transform: translateY(-50%);
+  }
+}
+
+@keyframes bottom-drip {
+  0% {
+    left: 45%;
+    transform: translateY(-100%);
+    width: 10%;
+  }
+  5% {
+    transform: translateY(0);
+  }
+  45%,
+  100% {
+    left: 50%;
+    transform: translateY(0);
+    width: 0;
+  }
+}
+
+@keyframes top-drip {
+  0%,
+  50% {
+    left: 45%;
+    transform: translateY(100%);
+    width: 10%;
+  }
+  55% {
+    left: 45%;
+    transform: translateY(0);
+    width: 10%;
+  }
+  100% {
+    left: 50%;
+    transform: translateY(0);
+    width: 0;
+  }
+}
